Fixed: 64bit issues, encapsulation issues, memory leaks, makefile commands.
64b issues: int cannot be used for dimensions passed to mex commands on 64bit systems. size_t can and should be valid also for 32bit.
encapsulation: in release more, CKernel pkern is private in the CGp class. So I made a const getter to access it.
memory leaks: unnecessary clone()-s for CTransform.
makefile: added commands to compile objects for CGp, CIvm & ndlstrutil to give them all necessary flags.

+++ CGp.cpp
@@ -1435,7 +1435,7 @@ CGp::CGp(CMatrix* pinData,
 }
 mxArray* CGp::toMxArray() const
 {
-  int dims[1];
+  size_t dims[1];
   dims[0]=1;
   mxArray* matlabArray;
   if(isSparseApproximation())
