+++ CNoise.cpp
@@ -59,7 +59,7 @@ void CNoise::updateSites(CMatrix& m, CMatrix& beta, unsigned int actIndex,
 #ifdef _NDLMATLAB
 mxArray* CNoise::toMxArray() const
 {
-  int dims[1];
+  size_t dims[1];
   dims[0] = 1;
   const char *fieldNames[] = {"type", "transforms", "numProcess", "spherical", "nParams", "missing", "logconcave"};
